Move the event runner to another package
